add assertion when payment is aborted
